+++ admin/frontend/components/sidebar.js
@@ -10,8 +10,10 @@ import Organizations from '../pages/organizations/list'
 import Roles from '../pages/roles/list'
 import Groups from '../pages/groups/list'
 
-import RequestLogs from '../pages/request-logs/list'
+import RequestLogs from '../pages/developer-tools/request-logs'
 import AppConfig from '../pages/developer-tools/app-config'
+import FormBuilder from '../pages/developer-tools/form-builder'
+// #Import
 
 class Sidebar extends Component {
   constructor (props) {
@@ -51,12 +53,13 @@ class Sidebar extends Component {
         open: false,
         dropdown: [
           Users.asSidebarItem(),
-          DeletedUsers.asSidebarItem(),
           Organizations.asSidebarItem(),
           Roles.asSidebarItem(),
           Groups.asSidebarItem()
         ]
-      }, {
+      },
+      // #Modules
+      {
         title: 'Load Data',
         icon: 'file-o',
         to: '/import',
@@ -71,7 +74,16 @@ class Sidebar extends Component {
         open: false,
         dropdown: [
           RequestLogs.asSidebarItem(),
-          AppConfig.asSidebarItem()
+          AppConfig.asSidebarItem(),
+          FormBuilder.asSidebarItem()
+        ]
+      }, {
+        title: 'Restore data',
+        icon: 'trash-o',
+        to: '/restore',
+        open: false,
+        dropdown: [
+          DeletedUsers.asSidebarItem()// #Restore
         ]
       }
     ]