+++ app/frontend/components/layout.js
@@ -23,11 +23,6 @@ class Layout extends Component {
       this.setState({user})
     })
 
-    const config = await api.get('/app-config')
-    tree.set('config', config)
-
-    tree.commit()
-
     var me
     if (tree.get('jwt')) {
       try {
@@ -48,6 +43,11 @@ class Layout extends Component {
 
       tree.commit()
     }
+
+    const config = await api.get('/app-config')
+    tree.set('config', config)
+
+    tree.commit()
 
     this.setState({loaded: true})
   }