+++ tasks/scaffolding/templates/model.js
@@ -2,7 +2,6 @@ const mongoose = require('mongoose')
 const { Schema } = require('mongoose')
 const { v4 } = require('uuid')
 const dataTables = require('mongoose-datatables')
-const moment = require('moment')
 
 const {{ name | lower }}Schema = new Schema({
   {% for item in fields -%}
@@ -12,9 +11,11 @@ const {{ name | lower }}Schema = new Schema({
       {{- item.name }}: { type: {{ item.type | capitalize}} {% if item.isRequired %}, required: {{ item.isRequired }} {% endif %} {% if item.default %} , default: '{{ item.default }}' {% endif %} },
     {%- endif %}
   {%- endfor %}
-  dateCreated: { type: Date, default: moment.utc },
+
   uuid: { type: String, default: v4 },
   isDeleted: { type: Boolean, default: false }
+}, {
+  timestamps: true
 })
 
 {{ name | lower }}Schema.plugin(dataTables)
@@ -25,8 +26,25 @@ const {{ name | lower }}Schema = new Schema({
     {% for item in fields -%}
       {{- item.name }}: this.{{ item.name }},
     {%- endfor %}
-    dateCreated: this.dateCreated
+    createdAt: this.createdAt
   }
 }
+
+{{ name | lower }}Schema.methods.toPublic = function () {
+  return {
+    uuid: this.uuid,
+    {% for item in fields -%}
+      {{- item.name }}: this.{{ item.name }},
+    {%- endfor %}
+    createdAt: this.createdAt
+  }
+}
+
+{{ name | lower }}Schema.plugin(dataTables, {
+  formatters: {
+    toAdmin: (item) => item.toAdmin(),
+    toPublic: (item) => item.toPublic()
+  }
+})
 
 module.exports = mongoose.model('{{ name | capitalize }}', {{ name | lower }}Schema)