+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.2] - 
+
+### Added
+- Added marble form component
+- Added form builder on developer tools
+
+### Changed
+- Improve csv exports using `json2csv` with loader and optional filename
+
+### Fixed
+- Fix list formatters on users, groups, roles and organizations
+- Reset password and invite users now use User tokens
+
 ## [0.6.1] - 2018-05-06
 
 ### Added